Split utils into separate files for testing purposes
